Move functions to connection store
