+++ packages/connections/src/components/connections.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/css';
-import React, { useCallback, useEffect, useReducer, useRef } from 'react';
+import React, { useCallback, useEffect } from 'react';
 import {
   MongoDBLogo,
   breakpoints,
@@ -7,25 +7,12 @@ import {
   spacing,
 } from '@mongodb-js/compass-components';
 import ConnectForm from '@mongodb-js/connect-form';
-import {
-  ConnectionInfo,
-  ConnectionStorage,
-  DataService,
-  getConnectionTitle,
-} from 'mongodb-data-service';
-import debugModule from 'debug';
+import { ConnectionInfo, DataService } from 'mongodb-data-service';
 
 import ResizableSidebar from './resizeable-sidebar';
 import FormHelp from './form-help/form-help';
-import { createConnectionAttempt } from '../modules/connection-attempt';
 import Connecting from './connecting/connecting';
-import {
-  connectionsReducer,
-  createNewConnectionInfo,
-  defaultConnectionsState,
-} from '../stores/connections-store';
-
-const debug = debugModule('mongodb-compass:connections:connections');
+import { useConnections } from '../stores/connections-store';
 
 const connectStyles = css({
   position: 'absolute',
@@ -79,118 +66,21 @@ function Connections({
       connections,
       isConnected,
     },
-    dispatch,
-  ] = useReducer(connectionsReducer, defaultConnectionsState());
-
-  const connectedConnectionInfo = useRef<ConnectionInfo>();
-  const connectedDataService = useRef<DataService>();
-
-  const updateActiveConnection = (newConnectionId?: string | undefined) => {
-    if (newConnectionId) {
-      const connection = connections.find(
-        (connection) => connection.id === newConnectionId
-      );
-      if (connection) {
-        dispatch({
-          type: 'set-active-connection',
-          connectionId: newConnectionId,
-          connectionInfo: connection,
-        });
-        return;
-      }
-    }
-
-    dispatch({
-      type: 'new-connection',
-      connectionInfo: createNewConnectionInfo(),
-    });
-  };
-
-  function onConnectSuccess(
-    dataService: DataService,
-    connectionInfo: ConnectionInfo
-  ) {
-    connectedConnectionInfo.current = connectionInfo;
-    connectedDataService.current = dataService;
-
-    dispatch({
-      type: 'connection-attempt-succeeded',
-    });
-    debug('connection attempt succeeded with connection info', connectionInfo);
-  }
-
-  async function onConnect(connectionInfo: ConnectionInfo) {
-    if (connectionAttempt || isConnected) {
-      // Ensure we aren't currently connecting.
-      return;
-    }
-
-    const newConnectionAttempt = createConnectionAttempt();
-
-    dispatch({
-      type: 'attempt-connect',
-      connectingStatusText: `Connecting to ${getConnectionTitle(
-        connectionInfo
-      )}`,
-      connectionAttempt: newConnectionAttempt,
-    });
-
-    debug('connecting with connectionInfo', connectionInfo);
-
-    try {
-      const connectedDataService = await newConnectionAttempt.connect(
-        connectionInfo.connectionOptions
-      );
-
-      if (!connectedDataService || newConnectionAttempt.isClosed()) {
-        // The connection attempt was cancelled.
-        return;
-      }
-
-      onConnectSuccess(connectedDataService, connectionInfo);
-    } catch (error) {
-      debug('connect error', error);
-
-      dispatch({
-        type: 'connection-attempt-errored',
-        connectionErrorMessage: error.message,
-      });
-    }
-  }
+    {
+      // state,
+      // cancelAnyCurrentConnectionAttempt,
+      loadConnections,
+      onCancelConnectionAttempt,
+      onConnect,
+      updateActiveConnection,
+    },
+  ] = useConnections(onConnected);
 
   const cancelAnyCurrentConnectionAttempt = useCallback(() => {
     if (!isConnected) {
       connectionAttempt?.cancelConnectionAttempt();
     }
   }, [connectionAttempt, isConnected]);
-
-  async function loadConnections() {
-    try {
-      const connectionStorage = new ConnectionStorage();
-      const loadedConnections = await connectionStorage.loadAll();
-
-      dispatch({
-        type: 'set-connections',
-        connections: loadedConnections,
-      });
-    } catch (error) {
-      debug('error loading connections', error);
-    }
-  }
-
-  useEffect(() => {
-    if (
-      isConnected &&
-      connectedConnectionInfo.current &&
-      connectedDataService.current
-    ) {
-      // After connecting and the UI is updated we notify the rest of Compass.
-      void onConnected(
-        connectedConnectionInfo.current,
-        connectedDataService.current
-      );
-    }
-  }, [isConnected, onConnected]);
 
   useEffect(() => {
     // Load connections after first render.
@@ -228,13 +118,7 @@ function Connections({
       {!!connectionAttempt && !connectionAttempt.isClosed() && (
         <Connecting
           connectingStatusText={connectingStatusText}
-          onCancelConnectionClicked={() => {
-            connectionAttempt?.cancelConnectionAttempt();
-
-            dispatch({
-              type: 'cancel-connection-attempt',
-            });
-          }}
+          onCancelConnectionClicked={onCancelConnectionAttempt}
         />
       )}
     </div>
